// src/app/models/sjcs/CertificacionFacItem.ts
export class CertificacionFacItem {

    fechaDesde: Date;
    fechaHasta: Date;
    idColegio: String;
    estado: String;
    idCodigo: String;
    idPartidaPresupuestaria: String;
    idFacturacion: String;
    idConcepto: String;
    nombre: String;
    periodo: String;
    importeTurno:String;
    importeGuardia:String;
    importeEjg:String;
    importeSoj:String;
    importeTotal:String;
    idCertificacion: String;
    fechaEstado: Date;
    proceso: String;
    grupoFacturacion: String;
    pendiente: String;
    facturado: String;
    regul: String;
    pagado: String;
    nColegiado: String;
    apellidos: String;
    nomColeg: String;
    descripcion: String;
    fechaAlta: Date;
    asunto: String;
    importeMovVario: String;

    constructor() { }
}